Moved sampling and smearing files to scn directory
