+++ ocpmodels/models/scn/smearing.py
